+++ Samples/MicroSvcsThreatsMapper/MicroSvcsThreatsMapper.csproj
@@ -87,7 +87,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ExtensionLibrary.cs" />
-    <Compile Include="MicroSvcsLoader.cs" />
+    <Compile Include="Libraries\ThreatModelLoader.cs" />
+    <Compile Include="Libraries\MicroSvcsLoader.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
